==> 0_download_data.ipynb <==
**[AgricultureVision](https://registry.opendata.aws/agriculture_vision/)**

Here, I explain how I downloaded the dataset associated with the [paper](https://arxiv.org/abs/2001.01306) [1] for my capstone project.

The complete dataset contains more than 94,000 high-resolution images from over three thousand farmlands in the US between 2017 and 2019. Each image was annotated with *double plant, drydown, endrow, nutrient deficiency, planter skip, storm damage, water, waterway and weed cluster*. Each image contains four channels: RGB and NIR. These annotated patterns or anomalies had a substancial impact in field conditions and subsequently crop yield [1].

I emailed the corresponding author, and asked if I could get access to the data. They very quickly, and shared with me a Google Drive link that I could download the data used in this [challenge](https://www.agriculture-vision.com/dataset) (~4.7 Gb). There is also more information on their [GitHub page](https://github.com/SHI-Labs/Agriculture-Vision) about this challenge, which was held at a workshop in June 2020. 

**I opted to download the data from the Google Drive link**, however I could have used AWS (below) to download the dataset from their [AWS S3 bucket](https://registry.opendata.aws/agriculture_vision/).


This subset dataset only includes the images from 2019, which contains over 21,000 farmland images and six annotatios: *cloud shadow, double plant, planter skip, standing water, waterway and weed cluster*. These six patterns are stored separately as binary masks due to potential overlaps between patterns. Each field image has a file name in the format of \(field id)_(x1)-(y1)-(x2)-(y2).(jpg/png). Each field id uniquely identifies the farmland that the image is cropped from, and (x1, y1, x2, y2) is a 4-tuple indicating the position in which the image is cropped.

mkdir dataset

cd dataset

%%bash
aws s3 ls s3://agriculture-vision/Dataset/

# option 1 - AWS
# 
# import subprocess
# files_to_download = ['data2017_miniscale.tar.gz', 'data2018_miniscale.tar.gz', 'data2019_miniscale.tar.gz',
#                      'data2017_splits.json', 'data2018_splits.json', 'data2019_splits.json']
# 
# for file in files_to_download:
#     cmd_to_run = f"aws s3 cp s3://agriculture-vision/Dataset/{file} ."
#     # print(cmd_to_run)
#     subprocess.call(cmd_to_run, shell=True)


# option 2 - Download from authors

```
$ gunzip Agriculture-Vision.tar.gz

$ tar -xvf Agriculture-Vision.tar
```

References:

1 - Mang Tik Chiu, Xingqian Xu, Yunchao Wei, Zilong Huang, Alexander Schwing, Robert Brunner,  Hrant Khachatrian, Hovnatan Karapetyan, Ivan Dozier, Greg Rose, David Wilson, Adrian Tudor, Naira Hovakimyan, Thomas S. Huang, and Honghui  Shi. Agriculture-vision: A large aerial image database for agricultural pattern analysis. arXiv preprint arXiv:2001.01306, 2020
