==> 1_install_libraries.ipynb <==
!pip install -U tensorflow

!pip install -U keras

!pip install opencv-python

!pip install natsort

import os
import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
import tensorflow as tf
import cv2

OK!

# optional stuff

!pip install jupyter_contrib_nbextensions

!pip install yapf

!pip install numpydoc

!pip install pyreadline

!pip install streamlit
