==> 3_select_single_images.ipynb <==
# Import libraries

%config Completer.use_jedi = False
import os
import pandas as pd
import numpy as np
import cv2
from PIL import Image
from shutil import copyfile

# Make new directory for single-labeled images

new_dir = 'partial_dataset'
train_images = os.path.join(new_dir, 'train', 'images') # I am using only the validation images

if not os.path.exists(new_dir):
    for d in [train_images]:
        os.makedirs(d)

# Save images in new directory

train_singles = pd.read_csv('validation_singles.csv')

list_labels = [('cloud_shadow', 1), ('double_plant', 2), ('planter_skip', 3), 
               ('standing_water',4), ('waterway', 5), ('weed_cluster', 6)]

for l in list_labels:
    train_singles[l[0]] = train_singles[l[0]].replace(1, l[1])
    
train_singles['filename'] = train_singles['filename'].str.split('/').str[-1].str[:-4]+'.jpg'
    
train_singles

# Save RGB image to `partial_dataset`

def save_images(df):    
    """ Save images. """
    
    for index, row in df.iterrows():
        filename = row['filename']
        img_source = os.path.join('dataset', 'Agriculture-Vision', 'val', 'images', 'rgb', filename)
        img_destination = os.path.join('partial_dataset', 'train', 'images', filename)
        copyfile(img_source, img_destination)

save_images(train_singles)

# Make labels as dataframe

labels = train_singles.set_index('filename').sum(axis=1).astype(int)
labels.to_csv(os.path.join('partial_dataset', 'train', 'labels_images.csv'))
