==> 4_TDI_CNN.ipynb <==
# Convolutional Neural Network (CNN)

I am using a CNN written in `tensorflow` and `keras`. CNN is a class of neural networks commonly applied to image data for image classification. In this technique, the image is divided into subgrids, and these subgrids are passed on the entire image. The passing of these subgrids is called convolution. In each convolution, the model is learning features that occupy that space at a higher level. In other words, in each convolution, the model passes the subgrids on the entire image and learns general features across the image.

# %%bash
# unzip /content/drive/MyDrive/TDI_Capstone_Project/partial_dataset.zipb

# Import libraries

%config Completer.use_jedi = False
import glob
import os
import cv2
from natsort import natsorted
from PIL import Image
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
import tensorflow as tf
from tensorflow import keras

print(f"Tensorflow version: {tf.__version__}")
print(f"Keras version: {keras.__version__}")

# Make labels

labels_images = pd.read_csv('partial_dataset/train/labels_images.csv')
labels_images.rename(columns={'0':'labels'}, inplace=True)
labels = labels_images['labels'].to_numpy() - 1  # to be in range(0,6)
labels

# Prepare images

# 1 - get list of paths
list_images = [f"partial_dataset/train/images/{x}" for x in labels_images['filename'].to_list()]
list_images[:5]

# 2 - function to read each image into np.array

def read_image(data_x):
    """Read obj with the path for images.
    data_x ==  *_x
    """

    W, H = 300, 300 # set final size

    x = cv2.imread(data_x, cv2.IMREAD_COLOR) # read in colors
    x = cv2.resize(x, (W, H)) # resize
    x = x / 255.0 # normalize values 0 to 1
    x = x.astype(np.float32) # make it np.float32
    
    return x

# 3 - get an array of images
images = np.array([read_image(x) for x in list_images])

# 4 - split labels and images into train and test
train_images, test_images = train_test_split(images, test_size=0.4, random_state=88)
train_labels, test_labels = train_test_split(labels, test_size=0.4, random_state=88)

print(
    f"Train images: {len(train_images)}, Train labels: {len(train_labels)}, Test images: {len(test_images)}, Test labels: {len(test_labels)}"
)

# check data type and shape
print(type(train_images), type(train_labels))
print(train_images.shape, train_labels.shape)

# check data type and shape
print(type(test_images), type(test_labels))
print(test_images.shape, test_labels.shape)

# Make a Convolutional Neural Network (CNN)

model = keras.Sequential([
        keras.layers.AveragePooling2D(pool_size=6, strides=6, input_shape=(300, 300, 3)),
        keras.layers.Conv2D(64, 3, activation='relu'),
        keras.layers.Conv2D(32, 2, activation='relu'),
        keras.layers.MaxPool2D(3, 3),
        keras.layers.Dropout(0.5),
        keras.layers.Flatten(),
        keras.layers.Dense(128, activation='relu'),
        keras.layers.Dense(6, activation='softmax')
])

model.compile(optimizer='adam',
              loss=keras.losses.SparseCategoricalCrossentropy(),
              metrics=['accuracy'])

model.fit(train_images, train_labels, epochs=10, batch_size=80)

acc = round(model.evaluate(test_images, test_labels, verbose=0)[1], 4)*100
print(f"Accuracy on test data is {acc}%")

model.save('./model_1_CNN')
